\name{abox}
\alias{abox}

\title{Draw a Thick Box Around a Plot}

\description{Draws a thick box (5 overlapped boxes) around a plot, in order to combat any anti-aliasing effects that may be in place.}

\usage{abox(...)}

\arguments{
  \item{...}{arguments to be passed to 'box'}
}

\author{Lee Kelvin <lee.kelvin@uibk.ac.at>}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
