\name{acol}
\alias{acol}

\title{Convert a named colour}

\description{Convert a named colour (or vector of colours) into their RGBA equivalent(s).}

\usage{acol(col, alpha = 0)}

\arguments{
  \item{col}{named colour (can be a vector)}
  \item{alpha}{alpha transparency value}
}

\value{
A vector equal in length to 'col' of RGBA equivalent colours.
}

\author{Lee Kelvin <lee.kelvin@uibk.ac.at>}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
