\name{astro-package}
\alias{astro-package}
\alias{astro}
\docType{package}
\title{
\packageTitle{astro}
}
\description{
\packageDescription{astro}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{astro}
\packageIndices{astro}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{astro}

Maintainer: \packageMaintainer{astro}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
