\name{combrad}
\alias{combrad}

\title{Combine Sersic Functions}

\description{Combine two or more Sersic functions and calculate the new implied total half-light radius}

\usage{combrad(mag, re, n)}

\arguments{
  \item{mag}{total magnitudes (lists of vectors)}
  \item{re}{half-light radii (lists of vectors)}
  \item{n}{Sersic indices (lists of vectors)}
}

\references{Graham A. W., Driver S. P., 2005, PASA, 22, 118}

\author{
Lee Kelvin, <lee.kelvin@uibk.ac.at>
}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}

