\name{convmu}
\alias{convmu}

\title{Convert Between Sersic Surface Brightnesses}

\description{Convert between Sersic surface brightnesses at different radii.}

\usage{convmu(r, mu, n, re = 1, rmu = re)}

\arguments{
  \item{r}{desired radius}
  \item{mu}{input surface brightness}
  \item{n}{Sersic index}
  \item{re}{half-light radius}
  \item{rmu}{radius at the input surface brightness}
}

\references{Graham A. W., Driver S. P., 2005, PASA, 22, 118}

\author{
Lee Kelvin, <lee.kelvin@uibk.ac.at>
}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
