\name{convrad}
\alias{convrad}

\title{Convert Between Sersic Radii}

\description{Convert between Sersic radii defined by different flux fractions.}

\usage{convrad(n, f = 0.9, r = 1, fr = 0.5)}

\arguments{
  \item{n}{Sersic index}
  \item{f}{flux fraction at new radius}
  \item{r}{reference radius}
  \item{fr}{flux fraction at reference radius}
}

\references{Graham A. W., Driver S. P., 2005, PASA, 22, 118}

\author{
Lee Kelvin, <lee.kelvin@uibk.ac.at>
}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
