\name{h2re}
\alias{h2re}

\title{Scalelength to Half-Light Radius}

\description{Convert scalelength to half-light radius.}

\usage{h2re(n, h = 1)}

\arguments{
  \item{n}{Sersic index}
  \item{h}{scalelength}
}

\references{Graham A. W., Driver S. P., 2005, PASA, 22, 118}

\author{
Lee Kelvin, <lee.kelvin@uibk.ac.at>
}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
