\name{igamma}
\alias{igamma}

\title{The Incomplete Gamma Function}

\description{Calculates and returns the incomplete gamma function (specifically, the lower incomplete gamma function).}

\usage{igamma(x, s)}

\arguments{
  \item{x}{upper limit of integration}
  \item{s}{shape parameter}
}

\references{Graham A. W., Driver S. P., 2005, PASA, 22, 118}

\author{
Lee Kelvin, <lee.kelvin@uibk.ac.at>
}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
