\name{nicetime}
\alias{nicetime}

\title{Convert Seconds into a Human Readable Time}

\description{Converts an input raw time value, such as that given by 'proc.time', and converts it into a human readable time format.}

\usage{nicetime(seconds)}

\arguments{
  \item{seconds}{number of seconds}
}

\value{
The output character string will display the input time (in seconds) in units of days, hours, minutes and seconds, as appropriate.
}

\author{Lee Kelvin <lee.kelvin@uibk.ac.at>}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
