\name{re2h}
\alias{re2h}

\title{Half-Light Radius to Scalelength}

\description{Convert half-light radius to scalelength.}

\usage{re2h(n, re = 1)}

\arguments{
  \item{n}{Sersic index}
  \item{re}{half-light radius}
}

\references{Graham A. W., Driver S. P., 2005, PASA, 22, 118}

\author{
Lee Kelvin, <lee.kelvin@uibk.ac.at>
}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
