\name{scmean}
\alias{scmean}

\title{Sigma-Clipped Mean}

\description{Calculates a sigma-clipped mean}

\usage{scmean(x, mult = 3, loop = 10)}

\arguments{
  \item{x}{a vector of data}
  \item{mult}{multiples of 1-sigma with which to clip the data (about the mean)}
  \item{loop}{number of loops clipping the data}
}

\author{Lee Kelvin <lee.kelvin@uibk.ac.at>}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
