\name{shade}
\alias{shade}

\title{Adds a Shaded Region to a Figure}

\description{A convenience wrapper around the 'polygon' function.}

\usage{shade(x, y1, y2, col = hsv(alpha=0.5), border = NA, ...)}

\arguments{
  \item{x}{x data}
  \item{y1}{y data (limit 1)}
  \item{y2}{y data (limit 2)}
  \item{col}{fill colour}
  \item{border}{border colour}
  \item{...}{arguments to be passed to 'polygon'}
}

\author{Lee Kelvin <lee.kelvin@uibk.ac.at>}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
