\name{solar}
\alias{solar}

\title{Display the Absolute Magnitude of the Sun}

\description{A quick reference for the absolute magnitude of the sun, in either AB or Vega.}

\usage{solar(band = "r", vega = FALSE, source = FALSE)}

\arguments{
  \item{band}{The band of interest (case sensitive)}
  \item{vega}{Output Vega magnitude instead?}
  \item{source}{Show the source of information?}
}

\author{Lee Kelvin <lee.kelvin@uibk.ac.at>}

\seealso{

The astronomy package: \code{\link{astro}}.

}

\keyword{data}
